refactor: 공통 상수 패키지 이동
