+++ src/main/java/lotto/model/LottoMachine.java
@@ -1,8 +1,8 @@
 package lotto.model;
 
-import static lotto.LottoNumberConstants.LOTTO_NUMBER_COUNT;
-import static lotto.LottoNumberConstants.LOTTO_NUMBER_MAX;
-import static lotto.LottoNumberConstants.LOTTO_NUMBER_MIN;
+import static lotto.constant.LottoNumberConstants.LOTTO_NUMBER_COUNT;
+import static lotto.constant.LottoNumberConstants.LOTTO_NUMBER_MAX;
+import static lotto.constant.LottoNumberConstants.LOTTO_NUMBER_MIN;
 
 import java.util.HashSet;
 import java.util.Random;
