+++ src/test/java/lotto/model/LottoTest.java
@@ -1,6 +1,6 @@
 package lotto.model;
 
-import static lotto.LottoNumberConstants.*;
+import static lotto.constant.LottoNumberConstants.*;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
